+++ src/main/java/insane96mcp/survivalreimagined/module/mining/blockhardness/BlockHardness.java
@@ -78,7 +78,7 @@ public class BlockHardness extends SRFeature {
 			new IdTagValue(IdTagMatcher.Type.ID, "minecraft:ancient_debris", 12d),
 			new IdTagValue(IdTagMatcher.Type.TAG, "survivalreimagined:obsidians", 33d),
 
-			new IdTagValue(IdTagMatcher.Type.ID, "minecraft:powder_snow", 1.5d)
+			new IdTagValue(IdTagMatcher.Type.ID, "minecraft:powder_snow", 2.5d)
 	));
 	public static final ArrayList<IdTagValue> customHardnesses = new ArrayList<>();