+++ changelog.md
@@ -6,6 +6,8 @@
   * As long as you're tired, the effect is applied, there's no longer a period between "You're tired and can sleep" and the effect
   * Effect is now applied indefinitely, and removed on wake up
   * Re-added Tiredness overlay (no longer shows up at Tired I)
+* Added back In terrain slowdown
+  * Snow and powder snow slow down
 
 ## 0.13.0
 * Update to MC 1.20.1